+++ NEWS.md
@@ -1,3 +1,52 @@
+# scDiffCom 1.0.0
+
+01-11-2023
+
+* Prepare for CRAN release
+* Update links to the Nature Aging paper
+
+# scDiffCom 0.2.4
+
+11-04-2023
+
+* Change default red-green colors in plots
+* Improve README and website
+
+# scDiffCom 0.2.3
+
+08-05-2022
+
+* Remove mouse and rat LRIs with "homolog_orthology_confidence" of 0, but only
+if they are not already provided directly by one of the "source database"
+* Remove some miss-curated LRIs
+
+# scDiffCom 0.2.2
+
+11-04-2022
+
+* Compute (permutation-based) p-values and logFC for each ligand and receptor gene independently
+
+# scDiffCom 0.2.1
+
+20-01-2022
+
+* Fix a bug that appeared when 'seurat_object[[]]' contained a column named 'cell_id'
+
+# scDiffCom 0.2.0
+
+24-11-2021
+
+* New function `BuildShiny()` that creates a shiny app from the results
+* New function `ReduceGO()` to reduce GO terms by semantic similarity
+
+# scDiffCom 0.1.0.9000
+
+09-11-2021
+
+* A rat LRI database (LRI_rat) has been added (based on orthology conversion)
+* `run_interaction_analysis()` accepts `rat` as species parameter
+* LRI datasets have a reduced size
+* Dataset `gene_ontology_level` is now available as data (instead of internal data)
 
 # scDiffCom 0.1